requirements.txt
django
psycopg2
djangorestframework
dj-database-url
gunicorn
django-heroku